third_party/WebKit/LayoutTests/fast/peerconnection/RTCPeerConnection-ice.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
</head>
<body>
<script>
function onIceChange2()
{
    if (window.pc.iceConnectionState === "closed") {
        calledCallbacks.push("iceConnectionState closed");
        assert_array_equals(calledCallbacks, [
          'iceConnectionState complete',
          'expectedTypeError',
          'expectedTypeError',
          'expectedTypeError',
          'expectedTypeError',
          'success1',
          'success2',
          'success3',
          'success4',
          'iceConnectionState closed'
        ]);
        window.testRTC.done();
    }
}

function addIceCandidateSuccess1()
{
    calledCallbacks.push('success1');
    window.pc.addIceCandidate({candidate: "candidate", sdpMid: 0}, addIceCandidateSuccess2, unexpectedCallback);
}

function addIceCandidateSuccess2()
{
    calledCallbacks.push('success2');
    window.pc.addIceCandidate({candidate: "candidate", sdpMLineIndex: 0}, addIceCandidateSuccess3, unexpectedCallback);
}

function addIceCandidateSuccess3()
{
    calledCallbacks.push('success3');
    window.pc.addIceCandidate({candidate: "candidate", sdpMid: 0, sdpMLineIndex: 0}, addIceCandidateSuccess4, unexpectedCallback);
}

function addIceCandidateSuccess4()
{
    calledCallbacks.push('success4');
    window.pc.oniceconnectionstatechange = onIceChange2;
    window.pc.close();
}

function unexpectedCallback()
{
    calledCallbacks.push('unexpected');
}

function expectedTypeError(error)
{
  assert_equals(error.name, 'TypeError');
  calledCallbacks.push('expectedTypeError');
}

function onIceChange1()
{
    if (window.pc.iceConnectionState === "completed") {
        calledCallbacks.push('iceConnectionState complete');
        var iceCandidate = new RTCIceCandidate({candidate: "nano nano"});
        window.pc.addIceCandidate(null, unexpectedCallback, unexpectedCallback).catch(expectedTypeError);
        window.pc.addIceCandidate({candidate: "candidate"}, unexpectedCallback, unexpectedCallback).catch(expectedTypeError);
        window.pc.addIceCandidate(iceCandidate, null, unexpectedCallback).catch(expectedTypeError);
        window.pc.addIceCandidate(iceCandidate, unexpectedCallback, null).catch(expectedTypeError);
        window.pc.addIceCandidate(iceCandidate, addIceCandidateSuccess1, unexpectedCallback);
    }
}

function testExecutionOrderClosedConnection()
{
    var localPeerConnection = new webkitRTCPeerConnection(null, null);
    localPeerConnection.close();
    var counter = 0;
    window.events = [];
    Promise.resolve().then(_ => events[counter++] = 1);
    var iceCandidate = new RTCIceCandidate({candidate:"nano nano"});
    localPeerConnection.addIceCandidate(iceCandidate, unexpectedCallback, error => {
        assert_equals(error.name, 'InvalidStateError');
        assert_equals(error.toString(), 'InvalidStateError: The RTCPeerConnection\'s signalingState is \'closed\'.');
        events[counter++] = 2;
    });
    Promise.resolve().then(_ => {
        events[counter++] = 3;
        assert_array_equals(events, [1, 2, 3]);
    });
}

window.calledCallbacks = [];
window.testRTC = async_test("Tests the RTCPeerConnection Ice functionality.");

testExecutionOrderClosedConnection();
window.pc = new webkitRTCPeerConnection(null, null);
window.pc.oniceconnectionstatechange = onIceChange1;
</script>
</body>
</html>
